% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_data_table}
\alias{to_data_table}
\title{Check and convert data to data.table}
\usage{
to_data_table(data)
}
\arguments{
\item{data}{A data.frame or data.table}
}
\value{
A data.table
}
\description{
Check and convert data to data.table
}
